+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/scene/interactive_scene.py
@@ -19,7 +19,7 @@ from omni.isaac.version import get_version
 from pxr import PhysxSchema
 
 from omni.isaac.orbit.assets import Articulation, ArticulationCfg, AssetBaseCfg, RigidObject, RigidObjectCfg
-from omni.isaac.orbit.sensors import SensorBase, SensorBaseCfg
+from omni.isaac.orbit.sensors import FrameTransformerCfg, SensorBase, SensorBaseCfg
 from omni.isaac.orbit.terrains import TerrainImporter, TerrainImporterCfg
 
 from .interactive_scene_cfg import InteractiveSceneCfg
@@ -342,6 +342,14 @@ class InteractiveScene:
             elif isinstance(asset_cfg, RigidObjectCfg):
                 self.rigid_objects[asset_name] = asset_cfg.class_type(asset_cfg)
             elif isinstance(asset_cfg, SensorBaseCfg):
+                # Update target frame path(s)' regex name space for FrameTransformer
+                if isinstance(asset_cfg, FrameTransformerCfg):
+                    updated_target_frames = []
+                    for target_frame in asset_cfg.target_frames:
+                        target_frame.prim_path = target_frame.prim_path.format(ENV_REGEX_NS=self.env_regex_ns)
+                        updated_target_frames.append(target_frame)
+                    asset_cfg.target_frames = updated_target_frames
+
                 self.sensors[asset_name] = asset_cfg.class_type(asset_cfg)
             elif isinstance(asset_cfg, AssetBaseCfg):
                 # manually spawn asset